2.3 - Attempt to fix errors. AUVAL passed.

+++ Source/Utility/SpectrumUtil.h
@@ -14,7 +14,7 @@ namespace SpectrumUtil
     inline float amp2db(float amp)
     {
         float dbNegative;
-        dbNegative = 20.0f*log10(amp);
+        dbNegative = 20.0f * std::log10(amp);
         
         return dbNegative;
     }
@@ -22,9 +22,9 @@ namespace SpectrumUtil
     /// fft bin to frequency conversion - probably won't use this one in spectrometer
     inline float bin2freq(float sr, uint32_t maxBin, uint32_t bin)
     {
-        float freqInterval = sr/(maxBin-1);
+        float freqInterval = sr / static_cast<float>(maxBin - 1);
         float frequency;
-        frequency = freqInterval*bin;
+        frequency = freqInterval * static_cast<float>(bin);
         return frequency;
     }
 }
